Move status_codes to pip._internal.cli
