+++ src/pip/_internal/commands/configuration.py
@@ -3,10 +3,10 @@ import os
 import subprocess
 
 from pip._internal.cli.base_command import Command
+from pip._internal.cli.status_codes import ERROR, SUCCESS
 from pip._internal.configuration import Configuration, kinds
 from pip._internal.exceptions import PipError
 from pip._internal.locations import venv_config_file
-from pip._internal.status_codes import ERROR, SUCCESS
 from pip._internal.utils.misc import get_prog
 
 logger = logging.getLogger(__name__)
