+++ src/pip/_internal/commands/hash.py
@@ -5,7 +5,7 @@ import logging
 import sys
 
 from pip._internal.cli.base_command import Command
-from pip._internal.status_codes import ERROR
+from pip._internal.cli.status_codes import ERROR
 from pip._internal.utils.hashes import FAVORITE_HASH, STRONG_HASHES
 from pip._internal.utils.misc import read_chunks
