+++ src/pip/_internal/commands/search.py
@@ -16,7 +16,7 @@ from pip._internal.compat import get_terminal_size
 from pip._internal.download import PipXmlrpcTransport
 from pip._internal.exceptions import CommandError
 from pip._internal.models.index import PyPI
-from pip._internal.status_codes import NO_MATCHES_FOUND
+from pip._internal.cli.status_codes import NO_MATCHES_FOUND
 from pip._internal.utils.logging import indent_log
 
 logger = logging.getLogger(__name__)
