+++ tests/functional/test_configuration.py
@@ -5,7 +5,7 @@ import textwrap
 
 import pytest
 
-from pip._internal.status_codes import ERROR
+from pip._internal.cli.status_codes import ERROR
 from tests.lib.configuration_helpers import ConfigurationMixin, kinds
